+++ api/src/opentrons/protocols/geometry/module_geometry.py
@@ -16,7 +16,8 @@ from typing import (Mapping, Optional,
 
 import numpy as np  # type: ignore
 import jsonschema  # type: ignore
-from opentrons.protocols.implementations.labware import LabwareImplementation
+from opentrons.protocols.implementations.protocol_api.labware import\
+    LabwareImplementation
 
 from opentrons_shared_data import module
 from opentrons.types import Location, Point, LocationLabware
