+++ api/src/opentrons/protocols/labware/load.py
@@ -7,7 +7,8 @@ from opentrons.protocols.labware.definition import (
 from opentrons.calibration_storage import get
 from opentrons.protocols.implementations.interfaces.labware import \
     LabwareInterface
-from opentrons.protocols.implementations.labware import LabwareImplementation
+from opentrons.protocols.implementations.protocol_api.labware import \
+    LabwareImplementation
 from opentrons.types import Location
 from opentrons_shared_data.labware.dev_types import LabwareDefinition
