+++ api/src/opentrons/simulate.py
@@ -25,7 +25,7 @@ from opentrons.config import IS_ROBOT, JUPYTER_NOTEBOOK_LABWARE_DIR
 from opentrons import protocol_api
 from opentrons.commands import types as command_types
 from opentrons.protocols.api_support.util import HardwareToManage
-from opentrons.protocols.implementations.protocol_context import \
+from opentrons.protocols.implementations.protocol_api.protocol_context import \
     ProtocolContextImplementation
 from opentrons.protocols import parse, bundle
 from opentrons.protocols.types import (
