+++ api/tests/opentrons/protocol_api/test_context.py
@@ -6,7 +6,7 @@ from unittest import mock
 import opentrons.protocol_api as papi
 import opentrons.protocols.api_support as papi_support
 import opentrons.protocols.geometry as papi_geometry
-from opentrons.protocols.implementations.protocol_context import \
+from opentrons.protocols.implementations.protocol_api.protocol_context import \
     ProtocolContextImplementation
 from opentrons.protocols.implementations.simulators.protocol_context import \
     ProtocolContextSimulation
