+++ api/tests/opentrons/protocols/execution/test_execute_json_v3.py
@@ -1,5 +1,6 @@
 from opentrons.protocols.geometry.well_geometry import WellGeometry
-from opentrons.protocols.implementations.labware import LabwareImplementation
+from opentrons.protocols.implementations.protocol_api.labware import\
+    LabwareImplementation
 from opentrons.protocols.implementations.well import WellImplementation
 
 from unittest import mock
