+++ robot-server/tests/conftest.py
@@ -14,7 +14,8 @@ from fastapi import routing
 import pytest
 from datetime import datetime
 
-from opentrons.protocols.implementations.labware import LabwareImplementation
+from opentrons.protocols.implementations.protocol_api.labware import \
+    LabwareImplementation
 from starlette.testclient import TestClient
 from robot_server.constants import API_VERSION_HEADER, API_VERSION_LATEST
 from robot_server.service.app import app
